Allow "weight: 0" in messages to mask them
Allow in message objects the additional key `weight`, which can be set
to 0 (or 1) to cause that message to be masked out (or left unmasked)
for training (similar to [1]). This is helpful for training the model to be robust and
capable of error recovery upon a bad assistant message.
A missing `weight` key defaults to weight 1, to guarantee downward compatibility.

Extend `tests/prompt_strategies/test_sharegpt.py` to contain messages with `weight` keys.

Extend `src/axolotl/prompters.py::_build_result` to return the turns with weights as additional tuple element.
Do this in axolotl directly instead of modifying `fastchat.conversation`'s `Conversation`.

Extend `src/axolotl/prompt_tokenizers.py::tokenize_prompt` to mask out tokens when weight is set to 0.

[1]: https://github.com/mistralai/mistral-finetune